apps/telegrambot/requirements.txt
python-dotenv>=1.0.0
google-genai>=1.0.0
python-telegram-bot[job-queue]>=20.0
aiofiles>=23.2.0
pytz>=2023.3
aiohttp>=3.0.0
nest_asyncio>=1.5.0
sniffio>=1.3.0
httpx>=0.27.0
tiktoken>=0.5.2
pytest>=7.0.0
python-json-logger>=2.0.0
pydantic-settings>=2.3.4
